# doc/api.yaml
...
openapi: 3.0.3 
info:
  title: WASAPhoto backend REST APIs
  description: |-
    Keep in touch with your friends by sharing photos of special moments, 
    thanks to WASAPhoto! You can upload your photos directly from your PC, 
    and they will be visible to everyone following you.
  version: "0.0.1"
components:
  securitySchemes:
    bearerAuth:
      description: API authentication to authorize requests
      type: http
      scheme: bearer
  responses:
    BadRequest: 
      description: |-
        The request was not compliant with the documentation
        (eg. missing fields, etc).
    InternalServerError: 
      description: |-
        The server encountered an internal error.
        Further info in server logs.
    NotFound:
      description: Page not found.
...
    UnauthorizedError:
      description: Access token is missing or invalid.
  parameters:
    username:
      schema: {$ref: "#/components/schemas/username"}
      name: username
      in: path 
      required: true
    userFollow:
      schema: {$ref: "#/components/schemas/username"}
      name: userFollow
      in: path 
      required: true
...
    userLike:
      schema: {$ref: "#/components/schemas/username"}
      name: userLike
      in: path 
      required: true
    photo-id:
      schema:
        description: ID of a photo
        type: integer
        example: 1
      name: photo-id
      in: path
      required: true
...
  schemas:
...
    username:
      title: Username
      description: The username
      type: string
      pattern: |-
        ^(?=.{3,16}$)[a-zA-Z0-9._]+$
      minLength: 3 
      maxLength: 16
      example: Maria 
...
    id:
      description: Unique identifier.
      type: integer
      example: 1
      readOnly: true
    photo:
      title: Photo
      description: |-
        An object containing the photo file, its unique id, user that posted a caption
        and the timestamp.
      type: object
      properties:
        id: {$ref: "#/components/schemas/id"}
        photo-url: {$ref: "#/components/schemas/photoUrl"}
        timestamp: {$ref: "#/components/schemas/timestamp"}
        username: {$ref: "#/components/schemas/username"}
        caption: {$ref: "#/components/schemas/photoCaption"}
    photoCaption:
      description: The image's caption.
      type: string
      example: "Visiting CERN"
      pattern: ".*"
      minLength: 1
      maxLength: 250
    photoFile:
      type: string
      format: binary
      description: The photo file
      pattern: ".*"
      minLength: 0
      maxLength: 1000000
    photoUrl:
      type: string
      pattern: "http://.*"
      description: The url of the photo
      title: URL
      minLength: 10
      maxLength: 500
      example: http://localhost:8080/imageurl18087
...
    timestamp:
      title: Timestamp
      description: The timestamp 
      type: string
      minLength: 19
      maxLength: 19
      format: date-time
      pattern: |-
        [0-9]{1,4}-[0-9]{1,2}-[0-9]{1,2}T[0-9]{1,2}:[0-9]{1,2}:[0-9]{1,2}
      example: "2022-05-23T08:00:00"
      readOnly: true
...
paths:
...
  /users/{username}/media/: 
    parameters:
      - $ref: "#/components/parameters/username"
    post:
      security:
        - bearerAuth: []
      tags: ["media"]
      summary: Upload a new photo
      description: |-
        Adds a photo to the profile in path if matches the one authenticated.
      operationId: uploadPhoto
      requestBody:
        description: Photo details
        content:
          multipart/form-data:
            schema:
              description: An object containing the photo file and its caption
              type: object
              properties:
                photoCaption: {$ref: "#/components/schemas/photoCaption"}
                photoFile: {$ref: "#/components/schemas/photoFile"}
        required: true

      responses:
        "201":
          description: photo uploaded successfully.
          content:
            application/json:
              schema: {$ref: "#/components/schemas/photo"}
        "400": {$ref: "#/components/responses/BadRequest"}
        '401': {$ref: '#/components/responses/UnauthorizedError'}
        "500": {$ref: "#/components/responses/InternalServerError"}
...
  /users/{username}/media/{photo-id}/likes/{userLike}/:
    parameters:
      - $ref: "#/components/parameters/username"
      - $ref: "#/components/parameters/photo-id"
      - $ref: "#/components/parameters/userLike"
    put:
      security:
        - bearerAuth: []
      tags: ["likes"]
      summary: Likes the photo
      description: |- 
        The user authenticated {userLike} sets a like to the photo with photo-id given, 
        posted by username {username} in path; if userLike is not banned by username.
      operationId: likePhoto
      responses:
        "201":
          description: Liked successfully.
          content:
            application/json:
              schema: {$ref: "#/components/schemas/photo"}
        "400": {$ref: "#/components/responses/BadRequest"}
        "401": {$ref: '#/components/responses/UnauthorizedError'}
        "404": {$ref: "#/components/responses/NotFound"}
        "500": {$ref: "#/components/responses/InternalServerError"}

    delete:
      security:
        - bearerAuth: []
      tags: ["likes"]
      summary: Unlikes the photo
      description: |-
        The user authenticated {userLike} if liked the photo with photo-id given, 
        posted by username in path {username}, removes the like; if userLike is not banned by username.
      operationId: unlikePhoto
      responses:
        "204":
          description: Unliked successfully.
        "400": {$ref: "#/components/responses/BadRequest"}
        "401": {$ref: '#/components/responses/UnauthorizedError'}
        "404": {$ref: "#/components/responses/NotFound"}
        "500": {$ref: "#/components/responses/InternalServerError"}
...
  /users/{username}/following/{userFollow}/:
    parameters:
      - $ref: "#/components/parameters/username"
      - $ref: "#/components/parameters/userFollow"
    put:
      security:
        - bearerAuth: []
      tags: ["follow"]
      summary: Follows a user
      description: |-
        Account authenticated {username} starts follow {userFollow}; if userFollow is not banned by username and username is not equal to userFollow.
      operationId: followUser
      responses:
        "201":
          description: Followed successfully
          content:
            application/json:
              schema: {$ref: "#/components/schemas/username"}
        "400": {$ref: "#/components/responses/BadRequest"}
        "401": {$ref: '#/components/responses/UnauthorizedError'}
        "404": {$ref: "#/components/responses/NotFound"}
        "500": {$ref: "#/components/responses/InternalServerError"}

    delete:
      security:
        - bearerAuth: []
      tags: ["follow"]
      summary: unfollow
      description: |-
        If the account authenticated {username} follows {userfolloW} then unfollows; if userFollow is not banned by username.
      operationId: unfollowUser
      responses:
        "204":
          description: Unfollowed successfully      
        "400": {$ref: "#/components/responses/BadRequest"}
        "401": {$ref: '#/components/responses/UnauthorizedError'}
        "404": {$ref: "#/components/responses/NotFound"}
        "500": {$ref: "#/components/responses/InternalServerError"}
...